requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai==1.30.1

# ── PDF parsing ───────────────────────────────────────────────────────────────
pymupdf==1.24.3          # fitz — text + image extraction
pdfplumber==0.11.0       # high-fidelity table extraction

# ── Vector store ─────────────────────────────────────────────────────────────
faiss-cpu==1.8.0
numpy==1.26.4

# ── Data / serialisation ─────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
pillow==10.3.0
tiktoken==0.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==8.3.0          # retry logic for API calls
aiofiles==23.2.1
